ChatBotDiscord/src/requirements.txt
PyNaCl==1.4.0
ai21 == 2.0.0
requests
pytz
duckdb
langchain-community
langchain-text-splitters
pandas